Expand dat/bombers cap & range info
Give endpoints of linear function, for both grass and concrete runways.
For dist<RANGE use CAPACITY; otherwise interpolate as
 MRCAP + (CAPACITY - MRCAP) * (dist - RANGE) / (MRANGE - RANGE)
CAPACITY is used for both grass and concrete; it is assumed that every
 aircraft can at least lift full bombload empty tanks on grass, even if
 the resulting RANGE is far shorter than any plausible raid.
All ranges are now in units of miles and include outbound only.
Game is not yet updated to consume these values yet.